Refactor todos custom hook
