+++ frontend/src/App.js
@@ -1,39 +1,15 @@
 import Header from "./components/Header";
 import styled from "styled-components/macro";
-import {useEffect, useState} from "react";
-import {deleteTodo, getTodos, postTodo, putTodo} from "./service/todo-api-service";
-import {getNextStatus} from "./service/todo-service";
+
 import NavigationBar from "./components/NavigationBar";
 import {BrowserRouter as Router, Route, Switch} from "react-router-dom";
 import Homepage from "./components/Homepage";
 import BoardPage from "./components/BoardPage";
+import useTodos from "./hooks/useTodos";
 
 function App() {
 
-    const [todos, setTodos] = useState([]);
-
-    const addTodo = (description) => {
-        postTodo(description)
-            .then(addedTodo => setTodos([...todos, addedTodo]))
-    }
-
-    const advanceTodo = (todo) => {
-        const newStatus = getNextStatus(todo.status)
-        const advancedTodo = {...todo, status: newStatus}
-        putTodo(advancedTodo)
-            .then(updatedTodo =>
-                setTodos(todos.map(item => updatedTodo.id === item.id ? advancedTodo : item)))
-    }
-
-    const removeTodo = (id) => {
-        deleteTodo(id)
-            .then(() => setTodos(todos.filter(todo => todo.id !== id)))
-    }
-
-    useEffect(() => {
-        getTodos()
-            .then(todos => setTodos(todos))
-    }, [])
+    const {todos, addTodo, advanceTodo, removeTodo} = useTodos()
 
     return (
         <Router>
